Add a chart repo server implementation for testing (#22)
It has been so hard to reliably integration-test chartify in CI as it depended on various real remote charts, just to ensure the code goes through helm-fetch and helm-dep-up invocations that is necessary to make sure chartify works with chart dependencies.

This is an attempt to resolve that issue, by introducing an implementation of Helm chart repository server for testing-purposes. It is an HTTP server that automatically `helm-package`s the local charts, generate the index.yaml, and serve the packages and the index via HTTP.

Any integration test can use this server like a standard remote chart repository server. The charts served by the server does not change except we change it ourselves, so that the test reliability issue is resolved.