Fix the lock-file-out-of-sync error with local chart + adhoc deps
It turns out we only removed unnecessary dependencies from Chart.yaml and requirements.yaml, and not from Chart.lock and requirements.lock. These dangling lock files seem to result in the mentioned error.

Since this change, any remote or local chart that went through the chartify process will have all the lock files omitted, so that the error is gone. This is fine because all the original and adhoc dependencies are already included in the chartify output chart, due to that it is generated by `helm template`ing the original chart with various tweaks.

Ref https://github.com/roboll/helmfile/issues/2074#issuecomment-1068335836